tools/coredump: allow running tests on darwin (#621)
