+++ .github/workflows/unit-test-on-pull-request.yml
@@ -137,6 +137,28 @@ jobs:
           name: integration-test-binaries-${{ matrix.target_arch }}
           path: support/*.test
 
+  coredump-test-macos:
+    name: Coredump tests (macOS)
+    runs-on: macos-latest
+    steps:
+      - name: Clone code
+        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+      - name: Set up Go
+        uses: actions/setup-go@d35c59abb061a4a6fb18e82ac0862c26744d6ab5 # v5.5.0
+        with:
+          go-version-file: go.mod
+          cache-dependency-path: go.sum
+      - name: Cache coredump modules
+        uses: actions/cache@5a3ec84eff668545956fd18022155c47e93e2684 # v4.2.3
+        with:
+          path: tools/coredump/modulecache
+          key: coredumps-arm64-${{ hashFiles('tools/coredump/testdata/*/*.json') }}
+          restore-keys: |
+            coredumps-arm64
+            coredumps-
+      - name: Run coredump tests
+        run: GODEBUG=asyncpreemptoff=1 go test -v ./tools/coredump/
+
   integration-tests:
     name: Integration tests (v${{ matrix.kernel }} ${{ matrix.target_arch }})
     runs-on: ubuntu-24.04
