+++ libpf/pfelf/machine_amd64.go
@@ -1,0 +1,10 @@
+//go:build amd64
+
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package pfelf // import "go.opentelemetry.io/ebpf-profiler/libpf/pfelf"
+
+import "debug/elf"
+
+const CurrentMachine = elf.EM_X86_64
