+++ process/debug_linux.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
